Use new config.ts for some config work (#42)
